fix: circular deps move: initializer models to application.ts
