+++ npm/react/libs/core/src/models/index.ts
@@ -1,2 +1,2 @@
-export { type ApplicationInitializer, InitializerError } from './initializer';
+export { type ApplicationInitializer, InitializerError } from './application';
 export type { Endpoint, Environment } from './environment';
